// lib/data/spaces.generated.ts
// AUTO-GENERATED
export interface CatalogSpace { id: number; name: string; slug: string; environment: string; image: string | null; }
export const catalogSpaces: CatalogSpace[] = [
  {
    "id": 1,
    "name": "Bathroom",
    "slug": "bathroom",
    "environment": "indoor",
    "image": "/categories/6.jpg"
  },
  {
    "id": 2,
    "name": "Kitchen",
    "slug": "kitchen",
    "environment": "indoor",
    "image": "/categories/4.jpg"
  },
  {
    "id": 3,
    "name": "Living Room",
    "slug": "living-room",
    "environment": "indoor",
    "image": "/categories/3.jpg"
  },
  {
    "id": 4,
    "name": "Bedroom",
    "slug": "bedroom",
    "environment": "indoor",
    "image": "/categories/5.jpg"
  },
  {
    "id": 5,
    "name": "Terrace",
    "slug": "terrace",
    "environment": "outdoor",
    "image": "/categories/7.jpg"
  },
  {
    "id": 6,
    "name": "Garden",
    "slug": "garden",
    "environment": "outdoor",
    "image": "/categories/8.jpg"
  },
  {
    "id": 7,
    "name": "Pool Area",
    "slug": "pool-area",
    "environment": "outdoor",
    "image": "/categories/9.jpg"
  },
  {
    "id": 8,
    "name": "Outdoor Kitchen",
    "slug": "outdoor-kitchen",
    "environment": "outdoor",
    "image": "/categories/10.jpg"
  }
];